srgan_implementation/requirements.txt
torch==2.0.1
torchvision==0.15.2
torchaudio==2.0.2
albumentations==1.3.0
numpy==1.23.5
Pillow==10.0.0
tqdm==4.66.1